refactor(mjh): extract Profile display helpers (file-size no-growth)
Profile.tsx was already over the 500-LOC growth guard (553) and the
formatDateRange change grew it to 556. Extract REMOTE_PREF_LABELS +
formatDateRange to features/profile/profile-format.ts (556 -> 532).
