+++ apps/myjobhunter/frontend/src/pages/Profile.tsx
@@ -20,6 +20,7 @@ import ScreeningAnswerDialog from "@/features/profile/ScreeningAnswerDialog";
 import ResumeUploadSection from "@/features/profile/ResumeUploadSection";
 import SkillAddForm from "@/features/profile/SkillAddForm";
 import ScreeningAnswerRow from "@/features/profile/ScreeningAnswerRow";
+import { REMOTE_PREF_LABELS, formatDateRange } from "@/features/profile/profile-format";
 import { useGetProfileQuery, useUpdateProfileMutation } from "@/lib/profileApi";
 import { useListWorkHistoryQuery, useDeleteWorkHistoryMutation } from "@/lib/workHistoryApi";
 import { useListEducationQuery, useDeleteEducationMutation } from "@/lib/educationApi";
@@ -31,31 +32,6 @@ import {
 import type { WorkHistory } from "@/types/work-history/work-history";
 import type { Education } from "@/types/education/education";
 import type { ScreeningAnswer } from "@/types/screening-answer/screening-answer";
-
-// ---------------------------------------------------------------------------
-// Display helpers
-// ---------------------------------------------------------------------------
-
-const REMOTE_PREF_LABELS: Record<string, string> = {
-  remote_only: "Remote only",
-  hybrid: "Hybrid",
-  onsite: "On-site",
-  any: "Open to all",
-};
-
-function formatDateRange(start: string, end: string | null, isCurrent: boolean): string {
-  const fmt = (d: string) => {
-    const [year, month] = d.split("-");
-    return new Date(parseInt(year), parseInt(month) - 1).toLocaleDateString("en-US", {
-      year: "numeric",
-      month: "short",
-    });
-  };
-  if (isCurrent) {
-    return `${fmt(start)} – Present`;
-  }
-  return end ? `${fmt(start)} – ${fmt(end)}` : `${fmt(start)} – No end date`;
-}
 
 // ---------------------------------------------------------------------------
 // Main Profile page
